refactor: minor visual changes; moved handleObjectStateChange and handleFormStateChange to core/utils
