+++ frontend/src/features/auth/components/AuthBaseButton.js
@@ -2,8 +2,8 @@ import {Button} from "@mui/material";
 
 const AuthBaseButton = ({text, onClick}) => {
     return(
-        <Button variant="outlined" color="primary" size="large"
-                fullWidth onClick={onClick} sx={{borderRadius: "20px"}}>
+        <Button variant="outlined" size="large" fullWidth onClick={onClick}
+                sx={{borderRadius: "20px"}}>
             {text}
         </Button>
     );
