+++ frontend/src/features/auth/components/AuthMainGrid.js
@@ -2,7 +2,7 @@ import {Grid} from "@mui/material";
 
 const AuthMainGrid = ({content}) => {
     return(
-        <Grid container rowSpacing={4} direction="column"
+        <Grid container rowGap={4} direction="column"
               alignItems="center" justifyContent="center"
         >
             {content}
