+++ frontend/src/features/auth/utils.js
@@ -1,17 +1,3 @@
-export const handleObjectStateChange = (input, object, updateObject) => {
-    updateObject({
-        ...object,
-        ...input
-    });
-};
-
-export const handleFormStateChange = (e, form, updateForm) => {
-    updateForm({
-        ...form,
-        [e.target.name]: e.target.value.trim()
-    });
-};
-
 export const setTokensToLocalStorage = (response) => {
     localStorage.setItem('access_token', response.data.access_token);
     localStorage.setItem('refresh_token', response.data.refresh_token);
